Made it possible to work with filter-triggering speeches
Sometimes, AzureOpenAI will detect triggering in the data due to the presence of suspect hate speech/violent speech. This would previously break the `for` loop used to generate summaries for all relevant speeches, but is now fixed. In addition, we are now getting 4 additional columns containing metadata on language use related to hate, self-harm, sex and violence, out of which the first and the last one could be useful for data visualization.