Added proof-of-concept summary of party/personal views
Some things may need to be ironed out, and I need to run the code for everyone, but so far, the results look promising. I've also added proof-of-concept "human topics" for the LDA modelling, but that needs more work as the current implementation only delivers very short topic names (perhaps too short to be meaningful).