Added the ability to only import from new HTMLs
Previously, all HTML files in the `input/debatter` folder were imported and prepared for further use. Now, only newly added files will be read, and the processed data from the new files will then be appended to the previously cleaned data. This will ensure that we do not by accident overwrite old statements, which have their unique IDs, which are then used in the subsequent analyses.