Adjusted prompt for LDA-LLM topics & refreshed data
I adjusted the prompt used for the LLM-assisted LDA topic modelling, replacing the way too generic topic headlines with 1-2 sentences' summaries of the top 20% speeches for each topic. This will help create a more meaningful mapping and grouping of the topics into the correct categories, and will reduce confusion. The mapping table is to be updated separately.