Minor fixes to some charts
Mostly Fig 9 and 10 where some logical errors related to filtering/the variable shown on the chart were present. I still need to go through all charts with a legend and ensure that the series are ordered in the same way in both the DK and EN versions of the charts.